Factor out portForwardPodWithPrefix()
